refactor: move find_csv_files to config module and rename test
- Move find_csv_files from pipeline.rs to config.rs where its only
  caller (into_tasks) lives, aligning dependency direction
- Rename config_build_file_auto_names to
  config_build_file_defers_name_resolution to reflect current behavior
